+++ backend/src/main/java/calculator/functions/Arcsinus.java
@@ -12,7 +12,7 @@ import ch.obermuhlner.math.big.BigDecimalMath;
 
 /**
  * This class represents the arithmetic unary operation "asin".
- * The class extends an abstract superclass UnaryOperation.
+ * The class extends an abstract superclass UnaryFunction.
  */
 public final class Arcsinus extends UnaryFunction {