+++ CHANGELOG.md
@@ -10,8 +10,12 @@ All notable changes to this project will be documented in this file.
 - Support for Unary operations
 - Trigonometric functions: sine, cosine, tangent and their inverses.
 - Others functions: logarithm, square root.
+<<<<<<< Nicolas/FEAT/API
 - Power operator
 - REST API
+=======
+- Command-line interface (CLI): commmand help, exit and eval.-
+>>>>>>> dev
 
 ### Changed
 - update dependencies (thanks to dependabot)